requirements.txt
# Core Dependencies
appium-python-client>=2.15.0
PyYAML>=6.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

# Image Processing
Pillow>=10.0.0  # For image manipulation

# Data Handling
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-html>=4.0.0
pytest-cov>=4.0.0

# Code Quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Development
ipython>=8.0.0
jupyter>=1.0.0

# Documentation
sphinx>=7.0.0
sphinx-rtd-theme>=1.2.0
